# src/docs/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Organize API",
    "description": "Api de reconhecimento facial",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/usuarios/": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "description": "Retorna lista de registros",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna lista de registros",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "usuario_id": {
                    "type": "number",
                    "example": 1
                  },
                  "nome": {
                    "type": "string",
                    "example": "texto"
                  },
                  "email": {
                    "type": "string",
                    "example": "texto"
                  },
                  "senha": {
                    "type": "string",
                    "example": "texto"
                  },
                  "status": {
                    "type": "number",
                    "example": 1
                  },
                  "departamento_id": {
                    "type": "number",
                    "example": 1
                  },
                  "created_at": {
                    "type": "string",
                    "example": "data"
                  },
                  "updated_at": {
                    "type": "string",
                    "example": "data"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Erro interno",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "error"
                },
                "description": {
                  "type": "string",
                  "example": "mensagem do sistema"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Usuarios"
        ],
        "description": "Cria um registro",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "texto"
                },
                "email": {
                  "type": "string",
                  "example": "texto"
                },
                "senha": {
                  "type": "string",
                  "example": "texto"
                },
                "status": {
                  "type": "number",
                  "example": 1
                },
                "departamento_id": {
                  "type": "number",
                  "example": 1
                }
              },
              "required": [
                "nome",
                "email",
                "senha",
                "status",
                "departamento_id"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registro criado com sucesso",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Registro criado com sucesso."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Erro interno",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "error"
                },
                "description": {
                  "type": "string",
                  "example": "mensagem do sistema"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/usuarios/{id}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "description": "Retorna um registro",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna um registro",
            "schema": {
              "type": "object",
              "properties": {
                "usuario_id": {
                  "type": "number",
                  "example": 1
                },
                "nome": {
                  "type": "string",
                  "example": "texto"
                },
                "email": {
                  "type": "string",
                  "example": "texto"
                },
                "senha": {
                  "type": "string",
                  "example": "texto"
                },
                "status": {
                  "type": "number",
                  "example": 1
                },
                "departamento_id": {
                  "type": "number",
                  "example": 1
                },
                "created_at": {
                  "type": "string",
                  "example": "data"
                },
                "updated_at": {
                  "type": "string",
                  "example": "data"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Erro interno",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "error"
                },
                "description": {
                  "type": "string",
                  "example": "mensagem do sistema"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Usuarios"
        ],
        "description": "Edita um registro",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "usuario_id": {
                  "type": "number",
                  "example": 1
                },
                "nome": {
                  "type": "string",
                  "example": "texto"
                },
                "email": {
                  "type": "string",
                  "example": "texto"
                },
                "senha": {
                  "type": "string",
                  "example": "texto"
                },
                "status": {
                  "type": "number",
                  "example": 1
                },
                "departamento_id": {
                  "type": "number",
                  "example": 1
                }
              },
              "required": [
                "usuario_id",
                "nome",
                "email",
                "senha",
                "status",
                "departamento_id"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registro atualizado",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Registro atualizado com sucesso."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Erro interno",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "error"
                },
                "description": {
                  "type": "string",
                  "example": "mensagem do sistema"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Usuarios"
        ],
        "description": "Deleta um registro",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Registro deletado com sucesso",
            "type": "success"
          },
          "422": {
            "description": "Erro interno",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "error"
                },
                "description": {
                  "type": "string",
                  "example": "mensagem do sistema"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}